Validate that the input to remember will not try to use the numeric %
The code in JsonSpec.remember assumes that the incoming argument will
be a string and uses the % operator to interpolate from the memory
hash into the string.  However, it is perfectly possible to pass in a
Numeric type (e.g. Fixnum) for which the % operator means something
totally different which causes the code to try to convert the hash to
fixnum, which it can't.

Fixes #90