Update cucumber.rb
Hello, I'm requesting this extra regex option get added to cucumber.rb to fix an issue I'm having with Scenario Outlines.

Here is an example:
```
Scenario Outline: Example of scenario outline issue
      Given User makes GET request to "www.give-me-a-json.com/api/json"
      And   User receives 200 response code
      Then  the JSON response at "data/someInteger" should be <integer>
      
      Examples:
      | integer |
      | 5       |
```
^^^ The above will give me an undefined step reference error while conversely.... 
```
Scenario: Example of scenario outline issue
      Given User makes GET request to "www.give-me-a-json.com/api/json"
      And   User receives 200 response code
      Then  the JSON response at "data/someInteger" should be 5
```
This will work fine. Adding <.*> will allow for examples table entries to be read, regardless of quotations. There may have to be an extra check within the method to account for someone trying to pass in a string without quotations using the examples table, which I'd be happy to provide if requested.